apply to ktor2 too
